+++ src/onegov.swissvotes/onegov/swissvotes/tests/test_views_vote.py
@@ -96,6 +96,10 @@ def test_view_vote(swissvotes_app):
             recommendations_other_yes="Pro Velo",
             recommendations_other_no=None,
             recommendations_other_free="Pro Natura, Greenpeace",
+            recommendations_divergent={
+                'fdp-fr_ch': 2,
+                'jcvp_ch': 2,
+            },
             national_council_election_year=1990,
             national_council_share_fdp=Decimal('01.10'),
             national_council_share_cvp=Decimal('02.10'),
@@ -180,6 +184,8 @@ def test_view_vote(swissvotes_app):
     assert "Pro Velo" in page
     assert "Pro Natura" in page
     assert "Greenpeace" in page
+    assert "FDP.Die Liberalen Frauen" in page
+    assert "Junge CVP" in page
     assert "22.2%" in page
     assert "Details" in page
     assert "Angenommen" in page